feat: Add configurable cache middleware
Add a cache middleware with configration option for namespace, storage engine and  default cache ttl
(maxAge). The middleware sits infornt of Fetch middleware and bypasses the fetch request if there is
valid cache available. Has implementation of localStorage and in-memory cache where the latter is
the default one. Defines interface for adding new cache providers and contains stub for redis. The
cache middleware can be assigned one or more buckets that each respond to one endpoint, optionally
be enabled/disabled and told to use specific cache key for data retrieval.